requirements.txt
flask
# tensorflow
numpy
pillow
opencv-python
matplotlib
seaborn
scikit-learn